backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic>=2.0
